HDDS-12350. Reduce duplication between OmBucketReadWrite tests
