+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/freon/TestOmBucketReadWriteFileOps.java
@@ -17,22 +17,20 @@
 
 package org.apache.hadoop.ozone.freon;
 
+import static org.apache.hadoop.ozone.freon.OmBucketTestUtils.verifyOMLockMetrics;
 import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_ADDRESS_KEY;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.net.URI;
 import java.util.Arrays;
 import java.util.List;
-import java.util.UUID;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.ozone.OzoneConsts;
 import org.apache.hadoop.ozone.TestDataUtil;
-import org.apache.hadoop.ozone.client.BucketArgs;
 import org.apache.hadoop.ozone.client.OzoneClient;
-import org.apache.hadoop.ozone.om.lock.OMLockMetrics;
+import org.apache.hadoop.ozone.freon.OmBucketTestUtils.ParameterBuilder;
 import org.apache.ozone.test.NonHATests;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -54,7 +52,8 @@ public abstract class TestOmBucketReadWriteFileOps implements NonHATests.TestCas
             .setNumOfWriteOperations(5)
             .setReadThreadPercentage(70)
             .setCountForRead(10)
-            .setCountForWrite(5),
+            .setCountForWrite(5)
+            .setDescription("default"),
         new ParameterBuilder()
             .setPrefixFilePath("/dir1/dir2/dir3")
             .setTotalThreadCount(10)
@@ -63,7 +62,8 @@ public abstract class TestOmBucketReadWriteFileOps implements NonHATests.TestCas
             .setReadThreadPercentage(80)
             .setBufferSize(128)
             .setCountForRead(10)
-            .setCountForWrite(5),
+            .setCountForWrite(5)
+            .setDescription("with increased buffer and read thread %"),
         new ParameterBuilder()
             .setPrefixFilePath("/")
             .setTotalThreadCount(15)
@@ -71,7 +71,8 @@ public abstract class TestOmBucketReadWriteFileOps implements NonHATests.TestCas
             .setNumOfWriteOperations(3)
             .setDataSize("128B")
             .setCountForRead(5)
-            .setCountForWrite(3),
+            .setCountForWrite(3)
+            .setDescription("with 128 byte object"),
         new ParameterBuilder()
             .setPrefixFilePath("/dir1/")
             .setTotalThreadCount(10)
@@ -80,13 +81,15 @@ public abstract class TestOmBucketReadWriteFileOps implements NonHATests.TestCas
             .setCountForRead(5)
             .setCountForWrite(3)
             .setDataSize("64B")
-            .setBufferSize(16),
+            .setBufferSize(16)
+            .setDescription("with 64 byte object"),
         new ParameterBuilder()
             .setPrefixFilePath("/dir1/dir2/dir3")
             .setTotalThreadCount(10)
             .setNumOfReadOperations(5)
             .setNumOfWriteOperations(0)
-            .setCountForRead(5),
+            .setCountForRead(5)
+            .setDescription("pure reads"),
         new ParameterBuilder()
             .setLength(64)
             .setPrefixFilePath("/dir1/dir2/dir3/dir4")
@@ -95,10 +98,11 @@ public abstract class TestOmBucketReadWriteFileOps implements NonHATests.TestCas
             .setNumOfWriteOperations(5)
             .setCountForRead(0)
             .setCountForWrite(5)
+            .setDescription("pure writes")
     );
   }
 
-  @ParameterizedTest
+  @ParameterizedTest(name = "{0}")
   @MethodSource("parameters")
   void testOmBucketReadWriteFileOps(ParameterBuilder parameterBuilder) throws Exception {
     try (OzoneClient client = cluster().newClient()) {
@@ -164,167 +168,5 @@ public abstract class TestOmBucketReadWriteFileOps implements NonHATests.TestCas
       }
     }
     assertEquals(expectedCount, actual, "Mismatch Count!");
-  }
-
-  private void verifyOMLockMetrics(OMLockMetrics omLockMetrics) {
-    String readLockWaitingTimeMsStat =
-        omLockMetrics.getReadLockWaitingTimeMsStat();
-    LOG.info("Read Lock Waiting Time Stat: " + readLockWaitingTimeMsStat);
-    LOG.info("Longest Read Lock Waiting Time (ms): " +
-        omLockMetrics.getLongestReadLockWaitingTimeMs());
-    int readWaitingSamples =
-        Integer.parseInt(readLockWaitingTimeMsStat.split(" ")[2]);
-    assertThat(readWaitingSamples).isPositive();
-
-    String readLockHeldTimeMsStat = omLockMetrics.getReadLockHeldTimeMsStat();
-    LOG.info("Read Lock Held Time Stat: " + readLockHeldTimeMsStat);
-    LOG.info("Longest Read Lock Held Time (ms): " +
-        omLockMetrics.getLongestReadLockHeldTimeMs());
-    int readHeldSamples =
-        Integer.parseInt(readLockHeldTimeMsStat.split(" ")[2]);
-    assertThat(readHeldSamples).isPositive();
-
-    String writeLockWaitingTimeMsStat =
-        omLockMetrics.getWriteLockWaitingTimeMsStat();
-    LOG.info("Write Lock Waiting Time Stat: " + writeLockWaitingTimeMsStat);
-    LOG.info("Longest Write Lock Waiting Time (ms): " +
-        omLockMetrics.getLongestWriteLockWaitingTimeMs());
-    int writeWaitingSamples =
-        Integer.parseInt(writeLockWaitingTimeMsStat.split(" ")[2]);
-    assertThat(writeWaitingSamples).isPositive();
-
-    String writeLockHeldTimeMsStat = omLockMetrics.getWriteLockHeldTimeMsStat();
-    LOG.info("Write Lock Held Time Stat: " + writeLockHeldTimeMsStat);
-    LOG.info("Longest Write Lock Held Time (ms): " +
-        omLockMetrics.getLongestWriteLockHeldTimeMs());
-    int writeHeldSamples =
-        Integer.parseInt(writeLockHeldTimeMsStat.split(" ")[2]);
-    assertThat(writeHeldSamples).isPositive();
-  }
-
-  static class ParameterBuilder {
-
-    private static final String BUCKET_NAME = "bucket1";
-
-    private final String volumeName = "vol-" + UUID.randomUUID();
-    private final BucketArgs.Builder bucketArgs = BucketArgs.newBuilder();
-    private String prefixFilePath = "/dir1/dir2";
-    private int countForRead = 100;
-    private int countForWrite = 10;
-    private String dataSize = "256B";
-    private int bufferSize = 64;
-    private int length = 10;
-    private int totalThreadCount = 100;
-    private int readThreadPercentage = 90;
-    private int numOfReadOperations = 50;
-    private int numOfWriteOperations = 10;
-
-    int getExpectedWriteCount() {
-      int readThreadCount = (getReadThreadPercentage() * getTotalThreadCount()) / 100;
-      int writeThreadCount = getTotalThreadCount() - readThreadCount;
-      return writeThreadCount * getCountForWrite() * getNumOfWriteOperations();
-    }
-
-    ParameterBuilder setPrefixFilePath(String newValue) {
-      prefixFilePath = newValue;
-      return this;
-    }
-
-    ParameterBuilder setCountForRead(int newValue) {
-      countForRead = newValue;
-      return this;
-    }
-
-    ParameterBuilder setCountForWrite(int newValue) {
-      countForWrite = newValue;
-      return this;
-    }
-
-    ParameterBuilder setDataSize(String newValue) {
-      dataSize = newValue;
-      return this;
-    }
-
-    ParameterBuilder setBufferSize(int newValue) {
-      bufferSize = newValue;
-      return this;
-    }
-
-    ParameterBuilder setLength(int newValue) {
-      length = newValue;
-      return this;
-    }
-
-    ParameterBuilder setTotalThreadCount(int newValue) {
-      totalThreadCount = newValue;
-      return this;
-    }
-
-    ParameterBuilder setReadThreadPercentage(int newValue) {
-      readThreadPercentage = newValue;
-      return this;
-    }
-
-    ParameterBuilder setNumOfReadOperations(int newValue) {
-      numOfReadOperations = newValue;
-      return this;
-    }
-
-    ParameterBuilder setNumOfWriteOperations(int newValue) {
-      numOfWriteOperations = newValue;
-      return this;
-    }
-
-    public String getVolumeName() {
-      return volumeName;
-    }
-
-    public String getBucketName() {
-      return BUCKET_NAME;
-    }
-
-    public String getPrefixFilePath() {
-      return prefixFilePath;
-    }
-
-    public BucketArgs.Builder getBucketArgs() {
-      return bucketArgs;
-    }
-
-    public int getBufferSize() {
-      return bufferSize;
-    }
-
-    public String getDataSize() {
-      return dataSize;
-    }
-
-    public int getCountForRead() {
-      return countForRead;
-    }
-
-    public int getCountForWrite() {
-      return countForWrite;
-    }
-
-    public int getLength() {
-      return length;
-    }
-
-    public int getTotalThreadCount() {
-      return totalThreadCount;
-    }
-
-    public int getReadThreadPercentage() {
-      return readThreadPercentage;
-    }
-
-    public int getNumOfReadOperations() {
-      return numOfReadOperations;
-    }
-
-    public int getNumOfWriteOperations() {
-      return numOfWriteOperations;
-    }
   }
 }
